Implemented Suggestions
- No msg is shown, if the file is missing or the description is empty
- Moved clsCustomDataDirectoryInfo into it's own file
- The Description is now a property
- Rewrote the manifest.xsd
