+++ Chummer/customdata/German Data Changes/manifest.xml
@@ -1,19 +1,29 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <manifest>
   <name>German Data Changes</name>
-  <guid>1ea1a21e-7cca-4ce9-b062-8f306880e8a6</guid>
   <!--version and author are just here to serve as an example-->
   <version>1.0</version>
-  <author>Multiple</author>
+  <authors>
+    <author>Author 1</author>
+    <author>Author 2</author>
+  </authors>
   <descriptions>
     <description lang="de-de">Dieser Custom Data Ordner versucht alle durch Pegasus in den Deutschen Regelwerken eingeführten Regeländerungen zu implementieren. Hierbei werden die Deutschen Bücher als errata betrachtet und überschreiben alte Objekte.</description>
     <description>This folder tries to implement all changes, that are made by Pegasus in the German Sourcebooks. Old English objects will be overridden by the new German rules.</description>
   </descriptions>
-  <!--Those are only present as example and should contain the GuID of other directories.-->
+  <!--Those are only present as example!-->
   <dependencies>
-    <dependency></dependency>
+    <dependency>
+      <name>Example Name</name>
+      <version>1</version>
+      <hash>I'm a hash, trust me</hash>
+    </dependency>
   </dependencies>
   <exclusivities>
-    <exclusivitie></exclusivitie>
+    <exclusivitie>
+      <name>Example Name 2</name>
+      <version>1</version>
+      <hash>I'm a hash, trust me</hash>
+    </exclusivitie>
   </exclusivities>
 </manifest>
