+++ Chummer/customdata/manifest.xsd
@@ -8,10 +8,15 @@
 >
   <xs:complexType name ="Manifest">
     <xs:sequence>
-      <xs:element name="guid" type="xs:string" minOccurs="1" maxOccurs="1"/>
       <xs:element name="name" type="xs:string" minOccurs="1" maxOccurs="1"/>
       <xs:element name="version" type="xs:string" minOccurs="0" maxOccurs="1"/>
-      <xs:element name="author" type="xs:string" minOccurs="0" maxOccurs="1"/>
+      <xs:element name="authors" type="xs:string" minOccurs="0" maxOccurs="1">
+        <xs:complexType>
+          <xs:sequence>
+            <xs:element name="author" type="xs:string" minOccurs="0" maxOccurs="unbounded"/>
+          </xs:sequence>
+        </xs:complexType>
+      </xs:element>
       <xs:element name="descriptions" minOccurs="0" maxOccurs="1">
         <xs:complexType>
           <xs:sequence>
@@ -23,17 +28,34 @@
       <xs:element name="dependencies" type="xs:string" minOccurs="0" maxOccurs="1">
         <xs:complexType>
           <xs:sequence>
-            <xs:element name="dependency" type="xs:string" minOccurs="0" maxOccurs="unbounded"/>
+            <xs:element name="dependency" type="xs:string" minOccurs="0" maxOccurs="unbounded">
+              <xs:complexType>
+                <xs:sequence>
+                  <xs:element name="name" type="xs:string" minOccurs="1" maxOccurs="1" />
+                  <xs:element name="version" type="xs:string" minOccurs="1" maxOccurs="1" />
+                  <xs:element name="hash" type="xs:string" minOccurs="1" maxOccurs="1" />
+                </xs:sequence>
+              </xs:complexType>
+            </xs:element>
           </xs:sequence>
         </xs:complexType>
       </xs:element>
       <xs:element name="exclusivities" type="xs:string" minOccurs="0" maxOccurs="1">
         <xs:complexType>
           <xs:sequence>
-            <xs:element name="exclusivity" type="xs:string" minOccurs="0" maxOccurs="unbounded"/>
+            <xs:element name="exclusivity" type="xs:string" minOccurs="0" maxOccurs="unbounded">
+              <xs:complexType>
+                <xs:sequence>
+                  <xs:element name="name" type="xs:string" minOccurs="1" maxOccurs="1" />
+                  <xs:element name="version" type="xs:string" minOccurs="1" maxOccurs="1" />
+                  <xs:element name="hash" type="xs:string" minOccurs="1" maxOccurs="1" />
+                </xs:sequence>
+              </xs:complexType>
+            </xs:element>
           </xs:sequence>
         </xs:complexType>
       </xs:element>
+      <xs:element name="hash" type="xs:string" minOccurs="0" maxOccurs="1"/>
     </xs:sequence>
   </xs:complexType>
 </xs:schema>
